requirements.txt
audioread==3.0.1
basic-pitch==0.3.3
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coloredlogs==15.0.1
decorator==5.1.1
flatbuffers==24.3.25
future==1.0.0
humanfriendly==10.0
idna==3.7
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.42.0
mido==1.3.2
mir_eval==0.7
mpmath==1.3.0
msgpack==1.0.8
mutagen==1.47.0
numba==0.59.1
numpy==1.26.4
onnxruntime==1.18.0
packaging==23.2
platformdirs==4.2.2
pooch==1.8.1
pretty_midi==0.2.10
protobuf==5.27.0
pycparser==2.22
pycryptodomex==3.20.0
pygame==2.5.2
pyo==1.0.5
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyreadline3==3.4.1
requests==2.32.2
resampy==0.4.2
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
threadpoolctl==3.5.0
typing_extensions==4.12.0
urllib3==2.2.1
websockets==12.0
yt-dlp==2024.4.9
